add distributive
змінив розташування файлу для збереження
прибрав аргументи для функцій main
зробив структуру для створення пакету
+ додав дистрибутив
