+++ bot/LICENSE
